changed a couple of variables to long longs
git-svn-id: svn://orca.physics.unc.edu/Orca@10583 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Misc Objects/NetSocket/ORSocketClient.h
@@ -26,7 +26,7 @@
     NetSocket*		socket;
     NSDate*			timeConnected;
     unsigned long	amountInBuffer;
-    unsigned long	totalSent;
+    unsigned long long	totalSent;
     int             socketStatus;
 }
 
@@ -37,8 +37,8 @@
 - (void) setDelegate:(id)newDelegate;
 - (NSString*) name;
 - (void) setName:(NSString*)newName;
-- (unsigned long) totalSent;
-- (void) setTotalSent:(unsigned long)aTotalSent;
+- (unsigned long long) totalSent;
+- (void) setTotalSent:(unsigned long long)aTotalSent;
 - (NetSocket*) socket;
 - (NSDate*) timeConnected;
 - (void) setTimeConnected:(NSDate*)newTimeConnected;
